+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -16,6 +16,9 @@
 - Hover text for complex relocation symbols contains incorrect multiplicative factors
 - Produce more specific diagnostic on grammar predicate failure
 
+#### Changed
+- Using LSP dynamic registration of file change notifications
+
 ## [1.12.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.11.1...1.12.0) (2024-03-05)
 
 #### Added